tests: integration: allow subdirectories in integration tests
