--- tests/tests/regression_fuzzer_tests/disabled/search_space_discontinuity.json
@@ -1,1 +1,0 @@
-{"error_type": "SCHEDULE", "code": 500, "error": "{\"type\":\"core:allowance_convergence\",\"cause\":\"INTERNAL\",\"errorType\":\"discontinuity\",\"message\":\"Failed to converge when computing allowances because of a discontinuity in the search space\",\"trace\":[{\"type\":\"allowance\",\"index\":\"1\"}]}", "infra_name": "small_infra", "path_payload": {"infra": 94, "name": "foo", "steps": [{"duration": 338.9396149098179, "waypoints": [{"track_section": "TA7", "offset": 144.01694825192175}]}, {"duration": 1, "waypoints": [{"track_section": "TB0", "offset": 669.2927737032946}]}]}, "schedule_payload": {"timetable": 74, "path": 16041, "schedules": [{"train_name": "foo", "labels": [], "departure_time": 0, "allowances": [{"allowance_type": "engineering", "begin_position": 655.7681522126178, "end_position": 2258.43307578361, "value": {"value_type": "percentage", "percentage": 8.101399265042275}, "capacity_speed_limit": 9.183335961543804, "distribution": "LINEAR"}, {"allowance_type": "engineering", "begin_position": 485.9086361515102, "end_position": 2286.516833610721, "value": {"value_type": "percentage", "percentage": 13.02719840932256}, "capacity_speed_limit": 3.4416022635242394, "distribution": "MARECO"}], "initial_speed": 0, "rolling_stock": "253"}]}}
